Updated the script to be directory and file based.
It turns out that as there is so much data involved (even in a small test case) it is best to place all the "basic" data into appropriate directories and files. The first part of this has been done as the dirs and files are created. The next work will be it iterate over these directories and create the KML files which will be used by the NetowrkLinks to load up more information about this {country-code:3}{geohash} section of the world